+++ test/exactly_lib_test/instructions/configuration/actor/z_package_suite.py
@@ -3,12 +3,13 @@ import unittest
 from exactly_lib.instructions.configuration import actor as sut
 from exactly_lib_test.common.help.test_resources.check_documentation import suite_for_instruction_documentation
 from exactly_lib_test.instructions.configuration.actor import failing_parse, command_line, file_interpreter, \
-    source_interpreter
+    source_interpreter, null
 
 
 def suite() -> unittest.TestSuite:
     return unittest.TestSuite([
         failing_parse.suite(),
+        null.suite(),
         command_line.suite(),
         file_interpreter.suite(),
         source_interpreter.suite(),